+++ src/test/converter2/issues/gh2430.js
@@ -1,0 +1,11 @@
+/**
+ * @param {object} props
+ * @param {any} props.value Value
+ * @param {any} props.trueValue True default
+ * @param {any} props.falseValue False default
+ */
+export const Checkbox = ({
+    value,
+    trueValue = true,
+    falseValue: renamed = false,
+}) => {};